Add files via upload
Imagine being able to enjoy a cold drink on a hot summer day or a warm cup of coffee on a chilly morning without needing electricity or access to a fridge. This is the goal of a new portable prototype that uses a solar panel and thermoelectric cooler (TEC) to gain and lose heat as needed. The TEC plays a crucial role in the project, and after careful consideration, the TEC1-12706 model was selected along with two heat sinks to distribute heat to containers made of PET plastic. Three 3.7V lithium batteries were used in series to provide 11.1V, and a battery management system (BMS) regulates charging and discharging. The solar panel was chosen to provide the necessary power for fast charging, and a 10W panel was used. The containers were designed with holes for the heat sink and water tap and surrounded by glass wool insulation to minimize heat loss. Overall, this portable prototype provides a sustainable and convenient solution for gaining and losing heat on the go.